+++ playwright.config.ts
@@ -1,5 +1,5 @@
 /**
- * MIT License, Copyright (c) 2023 - 2024 Heiko Lübbe
+ * GPLv3 License, Copyright (c) 2023 - 2024 Heiko Lübbe
  * WordPress plugin zitat-service, see https://github.com/muhme/quote_wordpress
  *
  * playwright.config.ts - Playwright configuration for E2E tests