+++ scripts/clean.sh
@@ -2,7 +2,7 @@
 #
 # scripts/clean.sh - delete all quote_wp_* docker containers
 #
-# MIT License, Copyright (c) 2023 - 2024 Heiko Lübbe
+# GPLv3 License, Copyright (c) 2023 - 2024 Heiko Lübbe
 # WordPress plugin zitat-service, see https://github.com/muhme/quote_wordpress
 
 echo '*** Removing all docker containers quote_wp_*'