+++ scripts/compose.sh
@@ -2,7 +2,7 @@
 #
 # scripts/compose.sh - delete the five docker containers and build them new
 #
-# MIT License, Copyright (c) 2023 - 2024 Heiko Lübbe
+# GPLv3 License, Copyright (c) 2023 - 2024 Heiko Lübbe
 # WordPress plugin zitat-service, see https://github.com/muhme/quote_wordpress
 
 NAME=quote_wp_