+++ scripts/pack.sh
@@ -2,7 +2,7 @@
 #
 # scripts/pack.sh - create the installable plugin zip
 #
-# MIT License, Copyright (c) 2023 - 2024 Heiko Lübbe
+# GPLv3 License, Copyright (c) 2023 - 2024 Heiko Lübbe
 # WordPress plugin zitat-service, see https://github.com/muhme/quote_wordpress
 
 # -e exist if a command fails