+++ scripts/test.sh
@@ -2,7 +2,7 @@
 #
 # scripts/test.sh - running end-to-end tests
 #
-# MIT License, Copyright (c) 2023 - 2024 Heiko Lübbe
+# GPLv3 License, Copyright (c) 2023 - 2024 Heiko Lübbe
 # WordPress plugin zitat-service, see https://github.com/muhme/quote_wordpress
 
 CONTAINER="quote_wp_wordpress"