+++ src/SelectControlCategory.js
@@ -1,7 +1,7 @@
 /**
  * src/selectControlCategory.js - block editor UI SelectControl implementation for categories
  *
- * MIT License, Copyright (c) 2023 - 2024 Heiko Lübbe
+ * GPLv3 License, Copyright (c) 2023 - 2024 Heiko Lübbe
  * WordPress plugin zitat-service, see https://github.com/muhme/quote_wordpress
  *
  */