+++ src/helper.php
@@ -2,7 +2,7 @@
 /**
  * src/helper.php - PHP constants and common used methods
  *
- * MIT License, Copyright (c) 2023 - 2024 Heiko Lübbe
+ * GPLv3 License, Copyright (c) 2023 - 2024 Heiko Lübbe
  * WordPress plugin zitat-service, see https://github.com/muhme/quote_wordpress
  *
  */