+++ test/plugin.logged.in.spec.ts
@@ -1,7 +1,7 @@
 /**
  * tests/plugin.logged.in.spec.ts - frontend E2E tests using the plugin
  *
- * MIT License, Copyright (c) 2023 - 2024 Heiko Lübbe
+ * GPLv3 License, Copyright (c) 2023 - 2024 Heiko Lübbe
  * WordPress plugin zitat-service, see https://github.com/muhme/quote_wordpress
  *
  */