+++ test/testHelper.ts
@@ -1,7 +1,7 @@
 /**
  * tests/testHelper.ts - utility methods used in the tests
  *
- * MIT License, Copyright (c) 2023 - 2024 Heiko Lübbe
+ * GPLv3 License, Copyright (c) 2023 - 2024 Heiko Lübbe
  * WordPress plugin zitat-service, see https://github.com/muhme/quote_wordpress
  *
  */